chatbot/requirements.txt
boto3>=1.40.7
duckdb>=1.3.2
fastapi[standard]>=0.116.1
langchain>=0.3.27
langchain-ollama>=0.3.6
langgraph>=0.6.4
numpy>=2.3.2
ollama>=0.5.3
pandas>=2.3.1
plotly>=6.2.0
pyarrow>=21.0.0
pydantic>=2.11.7
pydantic-settings>=2.10.1
pyyaml>=6.0.2
requests>=2.32.4
rich>=14.1.0
